# frontend/src/api/schema/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/scheduler/job/update": {
      "put": {
        "tags": [
          "scheduler-controller"
        ],
        "operationId": "updateJob",
        "parameters": [
          {
            "name": "cron",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerKeyDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/scheduler/trigger/resume": {
      "post": {
        "tags": [
          "scheduler-controller"
        ],
        "operationId": "resumeTrigger",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerKeyDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/scheduler/trigger/pause": {
      "post": {
        "tags": [
          "scheduler-controller"
        ],
        "operationId": "pauseTrigger",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerKeyDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/scheduler/job/trigger": {
      "post": {
        "tags": [
          "scheduler-controller"
        ],
        "operationId": "triggerJob",
        "parameters": [
          {
            "name": "startAt",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobKeyDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/scheduler/page-query": {
      "get": {
        "tags": [
          "scheduler-controller"
        ],
        "operationId": "pageQuery",
        "parameters": [
          {
            "name": "pageRequestDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PageRequestDto"
            }
          },
          {
            "name": "queryDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/QueryDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDtoListJobTriggerDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "TriggerKeyDto": {
        "required": [
          "group",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "group": {
            "type": "string"
          }
        }
      },
      "JobKeyDto": {
        "required": [
          "group",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "group": {
            "type": "string"
          }
        }
      },
...
      "PageRequestDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "sortBy": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sortFields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortFieldObject"
            }
          }
        }
      },
      "SortFieldObject": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "order": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC",
              "DEFAULT"
            ]
          }
        }
      },
      "QueryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "JobTriggerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "group": {
            "type": "string"
          },
          "className": {
            "type": "string"
          },
          "jobDataMap": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          },
          "triggerName": {
            "type": "string"
          },
          "triggerGroup": {
            "type": "string"
          },
          "schedulerType": {
            "type": "string"
          },
          "cronExpression": {
            "type": "string"
          },
          "startTime": {
            "type": "integer",
            "format": "int64"
          },
          "endTime": {
            "type": "integer",
            "format": "int64"
          },
          "nextFireTime": {
            "type": "integer",
            "format": "int64"
          },
          "previousFireTime": {
            "type": "integer",
            "format": "int64"
          },
          "triggerState": {
            "type": "string"
          },
          "triggerJobDataMap": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          }
        }
      },
      "PageResponseDtoListJobTriggerDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobTriggerDto"
            }
          }
        }
      },
...
    }
  }
}